Move Show Intercom to sauce
Moves Show Intercom functionality to the sauce directory. No
functionality changes.
