+++ sauce/features/general/show-intercom/settings.js
@@ -1,0 +1,8 @@
+module.exports = {
+  name: 'showIntercom',
+  type: 'checkbox',
+  default: false,
+  section: 'general',
+  title: 'Show Intercom',
+  description: "It's easy to just click 'X' on the intercom announcements when they show up and forget to read what was there. This feature adds an option to the account options popup (click your e-mail at the bottom left of the screen) to show the intercom again."
+};
