--- source/common/res/features/show-intercom/settings.json
@@ -1,13 +1,0 @@
-{
-         "name": "showIntercom",
-         "type": "checkbox",
-      "default": true,
-      "section": "general",
-        "title": "Show Intercom",
-  "description": "It's easy to just click 'X' on the intercom announcements when they show up and forget to read what was there. This feature add an option to the account-options popup (click your e-mail at the bottom left of the screen) to show the intercom again.",
-      "actions": {
-                    "true": [
-                      "injectScript", "main.js"
-                    ]
-                 }
-}
